Big commit: Updated and extended the login functionality
The returned user from Login() is a dictionary, so I had to convert it to an instance of Member in order to use its encapsulated methods. The arguments of the (Member) initialization method had to change to make the convertion happen. Besides that, I created a hardcoded interface method for members and admins, which represent the new Catalog methods (see and search for books).