Refactor add book to catalog functionality to admin class
Previously, the add book to catalog functionality was included in the catalog class it self. However, this functionality is only relevant to admins, so it makes more sense to move it to their class.
Overall, this change should make it easier to manage the codebase and ensure that only authorized users can add books to the catalog.