Refactor Person entity to remove unnecessary attributes
In this commit, I refactored the Person entity to remove the username, password, id and library_system attributes. Instead, we use the NationalInsuranceNumber (NIN) as the unique identifier for each person.

Since the NIN's should not be up for grab, the method that returns the NIN's for each system member became private. This method is still accessible within the Person class, but it cannot be accessed from outside the class.

By making these changes, we create a cleaner and more realistic design for the Person entity.