Refactoring Member and Admin class
It felt like search the catalog would be a method for the catalog class but after all it makes more sense to include in the member class, since it's the member that wants to search for a book in the catalog.
With this perspective I have moved a couple of other methods where this scenario occured. Besides I improved DRY principles by making 2 generic methods for getting data out of a json file and updating data in a json file.

Last but not least I created some #TODO's since I have added a collaborator to this git repo.