admin can lend a book item to a member
I instantiate another member for an already existing member. Due to this the library instances will not be the same. So, I have added a nasty workaround, but it's okay for now.

Besides that, I have also slightly improved the check_book_item_status_for_member() method because I did not cover the possibility of members that do not own book items yet.