add a list of books functionality
These changes allow administrators to import a list of new books to the system at once, using a json file. The method first prompts the user to input the name of the file to import, then checks if the file is in the correct format (json), and finally reads the data from the file using the "get_data" helper method. If the data is valid, the method iterates over each book in the list, validates their ISBN and adds them to the system using the "add_book" method. If a book's ISBN is invalid, then the book is not added. The method then prints a success message when books have been added.

In addition, I had to fix the add list of members and add book functions with minor changes in order to make this work.

Overall, this feature improves the user experience for administrators who need to add multiple books to the system, by allowing them to do so in a single step, using a common file format.