Fix mismatched user identities in loanItems caused by unique generation of national_insurance_number
Previously, users were experiencing incorrect identity assignments due to the unique regeneration of their national_insurance_number during user creation. This uniqueness was beneficial for user creation, but not for initialization during login, resulting in a mismatch between the 'Number' field in the data file and the 'insurance_number' in the instance. Consequently, erroneous member identities were associated with loanItems, leading to bugs in the loan process.