Update admin functionalities
Altering book items was not done appropriately. I treated a book item the same as a book which is incorrect. An edit book item functionality should allow one to modify the details or attributes of a physical copy (book item) of a book in the library. This could involve updating information such as metadata associated with that specific physical copy. And adding and deleting book items should only involve updating the paper copies.